complete Sign Up modal
