--- react-app/src/components/auth/SignUpForm.js
@@ -1,139 +1,0 @@
-import React, { useState } from "react";
-import { useDispatch } from "react-redux";
-import { Redirect } from 'react-router-dom';
-import { setUser } from "../../store/session"
-import { signUp } from '../../services/auth';
-
-const SignUpForm = ({authenticated, setAuthenticated}) => {
-  const dispatch = useDispatch();
-  const [errors, setErrors] = useState([]);
-  const [firstName, setFirstName] = useState("");
-  const [lastName, setLastName] = useState("");
-  const [email, setEmail] = useState("");
-  const [password, setPassword] = useState("");
-  const [repeatPassword, setRepeatPassword] = useState("");
-  const [imageFile, setImageFile] = useState(null);
-  const [location, setLocation] = useState("");
-
-  const onSignUp = async (e) => {
-    e.preventDefault();
-
-    if (password === repeatPassword) {
-      const user = await signUp(firstName, lastName, email, password, imageFile, location);
-      if (!user.errors) {
-        setAuthenticated(true);
-        dispatch(setUser(user));
-      } else {
-        setErrors(user.errors);
-      }
-    }
-  };
-
-  const updateFirstName = (e) => {
-    setFirstName(e.target.value);
-  };
-
-  const updateLastName = (e) => {
-    setLastName(e.target.value);
-  };
-
-  const updateEmail = (e) => {
-    setEmail(e.target.value);
-  };
-
-    const updatePassword = (e) => {
-      setPassword(e.target.value);
-    };
-
-    const updateRepeatPassword = (e) => {
-      setRepeatPassword(e.target.value);
-    };
-
-  const updateImageFile = (e) => {
-    setImageFile(e.target.files[0]);
-  };
-
-  const updateLocation = (e) => {
-    setLocation(e.target.value);
-  };
-
-  if (authenticated) {
-    return <Redirect to="/" />;
-  }
-
-  return (
-    <form onSubmit={onSignUp}>
-      <div>
-        {errors.map((error) => (
-          <div>{error}</div>
-        ))}
-      </div>
-      <div>
-        <label>First Name</label>
-        <input
-          type="text"
-          name="firstName"
-          onChange={updateFirstName}
-          value={firstName}
-        ></input>
-      </div>
-      <div>
-        <label>Last Name</label>
-        <input
-          type="text"
-          name="lastName"
-          onChange={updateLastName}
-          value={lastName}
-        ></input>
-      </div>
-      <div>
-        <label>Email</label>
-        <input
-          type="text"
-          name="email"
-          onChange={updateEmail}
-          value={email}
-        ></input>
-      </div>
-      <div>
-        <label>Location</label>
-        <input
-          type="text"
-          name="location"
-          onChange={updateLocation}
-          value={location}
-        ></input>
-      </div>
-      <div>
-        <label>Picture</label>
-        <input
-          type="file"
-          name="imageFile"
-          onChange={updateImageFile}
-        ></input>
-      </div>
-      <div>
-        <label>Password</label>
-        <input
-          type="password"
-          name="password"
-          onChange={updatePassword}
-          value={password}
-        ></input>
-      </div>
-      <div>
-        <label>Repeat Password</label>
-        <input
-          type="password"
-          name="repeat_password"
-          onChange={updateRepeatPassword}
-          value={repeatPassword}
-          required={true}
-        ></input>
-      </div>
-      <button type="submit">Sign Up</button>
-    </form>
-  );
-};
-
-export default SignUpForm;
